Moved login/logout to App.vue, replaced CustomerNavBar and AdminNavBar with AppBar
