+++ clients/webclient/src/components/Admin/AdminInvoicing.vue
@@ -1,8 +1,10 @@
 <template>
 	<v-container>
-		<v-row class="text-center">
-			<v-col cols="12">
-				<h3>AdminInvoicing-komponenten. Legge inn oversikt over abonnenter og fakturering?</h3>
+		<v-row>
+			<v-col>
+                <div v-for="sub in allSubscriptions" v-bind:key="sub.userId">
+                    <p class="font-weight-light"> {{sub.fullname}} </p>
+                </div>
 			</v-col>
 		</v-row>
 	</v-container>
@@ -11,7 +13,21 @@
 <script lang="ts">
 import Vue from 'vue';
 import Component from 'vue-class-component';
+import * as interfaces from "../../../../../server/src/interfaces";
+import { getVendorSubscriptions } from "../../api/api";
 
 @Component
-export default class AdminInvoicing extends Vue {}
+export default class AdminInvoicing extends Vue {
+    private allSubscriptions: interfaces.UserSubscription[] | null = null;
+
+    async created() {
+        const result = await getVendorSubscriptions();
+        console.log(result)
+        if (result != null) {
+            this.allSubscriptions = result;
+        }
+    }
+
+
+}
 </script>
