+++ clients/webclient/src/components/Users/CustomerProfile.vue
@@ -91,51 +91,51 @@
 <script lang="ts">
 import Vue from "vue";
 import Component from "vue-class-component";
-import { MenuItems } from "../../../../server/src/interfaces";
-import { getUserprofile, getUserSubscription } from "../api/api";
+import { MenuItems } from "../../../../../server/src/interfaces";
+import { getUserprofile, getUserSubscriptions } from "../../api/api";
 
 @Component
 export default class CustomerProfile extends Vue {
-  //Userprofile
-  private fullname = "";
-  private address = "";
-  private phone = "";
-  private email = "";
-  private allergies: string[] = [];
+    //Userprofile
+    private fullname = "";
+    private address = "";
+    private phone = "";
+    private email = "";
+    private allergies: string[] = [];
 
-  //Usersubscription
-  private approved = false;
-  private paused = false;
-  private schedule: MenuItems[] = [];
-  private noOfMeals = "";
-  private box = "";
+    //Usersubscription
+    private approved = false;
+    private paused = false;
+    private schedule: MenuItems[] = [];
+    private noOfMeals = "";
+    private box = "";
 
 
-async showUserProfile() {
-    let userRes = await getUserprofile();
-    if(userRes != null) {
-        this.fullname = userRes.fullname;
-        this.address = userRes.address;
-        this.phone = userRes.phone;
-        this.email = userRes.email;
-        this.allergies = userRes.allergies;
+    async showUserProfile() {
+        let userRes = await getUserprofile();
+        if(userRes != null) {
+            this.fullname = userRes.fullname;
+            this.address = userRes.address;
+            this.phone = userRes.phone;
+            this.email = userRes.email;
+            this.allergies = userRes.allergies;
+        }
+
+        //Må finnne en bedre måte å hente vendor på
+        let vendor = "lunsj@hjul.no";
+        let subscriptionRes = await getUserSubscriptions();
+        if(subscriptionRes != null) {
+        this.approved = subscriptionRes[0].approved;
+        this.paused = subscriptionRes[0].paused;
+        this.noOfMeals = subscriptionRes[0].noOfMeals.toString();
+        this.box = subscriptionRes[0].box;
+        this.schedule = subscriptionRes[0].schedule;
+        }
     }
 
-    //Må finnne en bedre måte å hente vendor på
-    let vendor = "lunsj@hjul.no";
-    let subscriptionRes = await getUserSubscription(vendor);
-    if(subscriptionRes != null) {
-      this.approved = subscriptionRes.approved;
-      this.paused = subscriptionRes.paused;
-      this.noOfMeals = subscriptionRes.noOfMeals.toString();
-      this.box = subscriptionRes.box;
-      this.schedule = subscriptionRes.schedule;
+    beforeMount() {
+        this.showUserProfile();
     }
-
-    }
-
-  beforeMount() {
-      this.showUserProfile();
-  }
 }
+
 </script>
